PHPC-2500: Skip errorLabels assignment for non-RuntimeException exceptions (#1974)
* PHPC-2500: Assert that property is declared before assigning in phongo_add_exception_prop()

* PHPC-2500: Use size_t for prop_len in phongo_add_exception_prop()

* PHPC-2500: Skip errorLabels assignment for non-RuntimeException exceptions

phongo_exception_add_error_labels() now checks that the current exception
is an instance of MongoDB\Driver\Exception\RuntimeException before
attempting to assign errorLabels. Exceptions such as
InvalidArgumentException do not inherit from RuntimeException and do not
declare the errorLabels property.

This replaces the ZEND_ASSERT in phongo_add_exception_prop() with a proper
runtime guard that is effective in both debug and release builds.

+++ src/phongo_error.h
@@ -29,7 +29,7 @@ typedef enum {
 	PHONGO_ERROR_LOGIC             = 9
 } phongo_error_domain_t;
 
-void              phongo_add_exception_prop(const char* prop, int prop_len, zval* value);
+void              phongo_add_exception_prop(const char* prop, size_t prop_len, zval* value);
 zend_class_entry* phongo_exception_from_mongoc_domain(mongoc_error_domain_t domain, mongoc_error_code_t code);
 zend_class_entry* phongo_exception_from_phongo_domain(phongo_error_domain_t domain);
 void              phongo_exception_add_error_labels(const bson_t* reply);
